remove deprecations
